pyproject.toml
[project]
name = "testpython"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "daphne>=4.2.0",
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
    "requests>=2.32.3",
]
